requirements.txt
Babel==2.6.0
Django==2.2.1
django-phonenumber-field==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
phonenumbers==8.10.12
pkg-resources==0.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
